homework app: split code into modules
