+++ src/apps/homework/api/serializers/answer.py
@@ -1,78 +1,12 @@
-from typing import TYPE_CHECKING
-
-from drf_spectacular.helpers import lazy_serializer
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
-from apps.homework.models import Answer, AnswerCrossCheck, AnswerImage, Question
-from apps.homework.models.reaction import Reaction
+from apps.homework.api.serializers.question import QuestionSerializer
+from apps.homework.api.serializers.reaction import ReactionDetailedSerializer
+from apps.homework.models import Answer, Question
+from apps.homework.models.answer import AnswerQuerySet
 from apps.users.api.serializers import UserSafeSerializer
 from core.serializers import MarkdownField, SoftField
-
-if TYPE_CHECKING:
-    from apps.homework.models.answer import AnswerQuerySet
-
-
-class ReactionDetailedSerializer(serializers.ModelSerializer):
-    author = UserSafeSerializer()
-    answer = serializers.CharField(source="answer.slug")
-
-    class Meta:
-        model = Reaction
-        fields = [
-            "slug",
-            "emoji",
-            "author",
-            "answer",
-        ]
-
-
-class ReactionCreateSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Reaction
-        fields = [
-            "emoji",
-            "slug",
-        ]
-
-
-class QuestionSerializer(serializers.ModelSerializer):
-    text = MarkdownField()
-
-    class Meta:
-        model = Question
-        fields = [
-            "slug",
-            "name",
-            "text",
-            "deadline",
-        ]
-
-
-class QuestionDetailSerializer(serializers.ModelSerializer):
-    text = MarkdownField()
-    breadcrumbs = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Question
-        fields = [
-            "breadcrumbs",
-            "slug",
-            "name",
-            "text",
-            "deadline",
-        ]
-
-    @extend_schema_field(lazy_serializer("apps.lms.api.serializers.BreadcrumbsSerializer")())
-    def get_breadcrumbs(self, question: Question) -> dict | None:
-        from apps.homework.breadcrumbs import get_lesson
-        from apps.lms.api.serializers import BreadcrumbsSerializer
-
-        lesson = get_lesson(question, user=self.context["request"].user)
-        if lesson is None:
-            return None
-
-        return BreadcrumbsSerializer(lesson).data
 
 
 class AnswerSerializer(serializers.ModelSerializer):
@@ -101,7 +35,7 @@ class AnswerSerializer(serializers.ModelSerializer):
             "reactions",
         ]
 
-    def get_descendants_queryset(self, answer: Answer) -> "AnswerQuerySet":
+    def get_descendants_queryset(self, answer: Answer) -> AnswerQuerySet:
         user = self.context["request"].user
         if user.has_perm("homework.see_all_answers"):
             return answer.get_comments()
@@ -189,18 +123,6 @@ class AnswerCommentTreeSerializer(AnswerTreeSerializer):
         ]
 
 
-class AnswerImageSerializer(serializers.ModelSerializer):
-    author = serializers.HiddenField(default=serializers.CurrentUserDefault())
-    image = serializers.FileField()
-
-    class Meta:
-        model = AnswerImage
-        fields = [
-            "author",
-            "image",
-        ]
-
-
 class AnswerSimpleSerializer(serializers.ModelSerializer):
     url = serializers.URLField(source="get_absolute_url")
     author = UserSafeSerializer()
@@ -209,18 +131,3 @@ class AnswerSimpleSerializer(serializers.ModelSerializer):
     class Meta:
         model = Answer
         fields = ("slug", "url", "author", "question")
-
-
-class AnswerCrossCheckSerializer(serializers.ModelSerializer):
-    answer = AnswerSimpleSerializer()
-    is_checked = serializers.SerializerMethodField()
-
-    class Meta:
-        model = AnswerCrossCheck
-        fields = (
-            "answer",
-            "is_checked",
-        )
-
-    def get_is_checked(self, obj: "AnswerCrossCheck") -> bool:
-        return obj.checked is not None
