+++ src/apps/homework/models/__init__.py
@@ -2,10 +2,12 @@ from apps.homework.models.answer import Answer
 from apps.homework.models.answer_cross_check import AnswerCrossCheck
 from apps.homework.models.answer_image import AnswerImage
 from apps.homework.models.question import Question
+from apps.homework.models.reaction import Reaction
 
 __all__ = [
     "Answer",
     "AnswerCrossCheck",
     "AnswerImage",
     "Question",
+    "Reaction",
 ]
