+++ src/apps/homework/urls.py
@@ -13,8 +13,8 @@ reaction_router.register("reactions", ReactionViewSet)
 urlpatterns = [
     path("questions/<uuid:slug>/", views.QuestionView.as_view()),
     path("comments/", views.AnswerCommentView.as_view()),
-    path("crosschecks/", views.AnswerCrossCheckView.as_view()),
-    path("answers/image/", views.AnswerImageUploadView.as_view()),
+    path("crosschecks/", views.CrossCheckView.as_view()),
+    path("answers/image/", views.ImageUploadView.as_view()),
     path("", include(router.urls)),
     path("answers/<uuid:answer_slug>/", include(reaction_router.urls)),
 ]
